requirements.txt
apispec==0.39.0
flake8==3.6.0
flake8-commas==2.0.0
flake8-quotes==1.0.0
flasgger==0.9.2
Flask==1.0.2
flask-cors==3.0.6
flask-jwt-extended==3.13.1
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
Flask-RESTful==0.3.7
gevent==1.4.0
gunicorn==19.9.0
marshmallow-sqlalchemy==0.14.1
psycopg2-binary==2.7.5
python-dotenv==0.9.1
